test(providers): isolate environment-proxy test in its own binary
